third_party/WebKit/LayoutTests/http/tests/serviceworker/unregister.html

<!DOCTYPE html>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script>
(function() {
    var t = async_test('Unregistering out of origin');
    navigator.serviceWorker.unregister('http://example.com/*').then(
        t.step_func(function() {
            assert_unreached('unregistering out of origin should fail');
        }),
        t.step_func(function(reason) {
            assert_equals(reason.name, 'SecurityError',
                          'unregistering out of origins should fail');
            t.done();
        }));
}());

(function() {
    var t = async_test('Unregistering with no existing registration');
    navigator.serviceWorker.unregister('/nothing-here/*').then(
        t.step_func(function(value) {
            assert_equals(value, undefined,
                          'unregistering with no existing registration ' +
                          'should succeed with no value');
            t.done();
        }),
        t.step_func(function(reason) {
            assert_unreached('unregistering with no existing registration ' +
                             'should not fail: ' + reason.name);
        }));
}());

(function() {
    var scope = '/register-then-unregister/*';

    var t = async_test('Register then unregister');
    navigator.serviceWorker.register(
        'resources/empty-worker.js',
        { scope: scope }
    ).then(
        function(worker) {
            return navigator.serviceWorker.unregister(scope);
        },
        t.step_func(function(reason) {
            assert_unreached('registration should succeed.');
        })
    ).then(
        function(value) {
            assert_equals(value, undefined,
                          'successful unregistration should succeed ' +
                          'with no value');
            t.done();
        },
        t.step_func(function(reason) {
            assert_unreached('unregistration should succeed.');
        })
    );
}());
</script>
